requirements.txt
django>=3.0.5
djangorestframework>=3.11
flake8>=3.6.0<3.7.0
psycopg2>=2.8.5